Move PO functionality to new i18n package
