+++ i18n/flows.go
@@ -1,0 +1,187 @@
+package i18n
+
+import (
+	"errors"
+	"fmt"
+	"sort"
+
+	"github.com/nyaruka/goflow/envs"
+	"github.com/nyaruka/goflow/flows"
+	"github.com/nyaruka/goflow/utils/dates"
+	"github.com/nyaruka/goflow/utils/uuids"
+)
+
+// TextLocation describes the location of a piece of extracted text
+type TextLocation struct {
+	Flow     flows.Flow
+	UUID     uuids.UUID
+	Property string
+	Index    int
+}
+
+type ExtractedText struct {
+	Locations   []TextLocation
+	Base        string
+	Translation string
+	Unique      bool
+}
+
+func ExtractFromFlows(lang envs.Language, excludeArgs bool, sources ...flows.Flow) (*PO, error) {
+	extracted, err := extractFromFlows(lang, excludeArgs, sources)
+	if err != nil {
+		return nil, err
+	}
+
+	merged := mergeExtracted(extracted)
+
+	return createPO(lang, merged), nil
+}
+
+func extractFromFlows(lang envs.Language, excludeArgs bool, sources []flows.Flow) ([]*ExtractedText, error) {
+	baseLanguage := envs.NilLanguage
+	extracted := make([]*ExtractedText, 0)
+	for _, flow := range sources {
+		if baseLanguage == envs.NilLanguage {
+			baseLanguage = flow.Language()
+		} else if baseLanguage != flow.Language() {
+			return nil, errors.New("flows use different base languages")
+		}
+
+		var targetTranslation flows.Translation
+		if lang != envs.NilLanguage {
+			targetTranslation = flow.Localization().GetTranslation(lang)
+		}
+
+		for _, node := range flow.Nodes() {
+			node.EnumerateLocalizedText(func(uuid uuids.UUID, property string, texts []string) {
+				if !excludeArgs || property != "arguments" {
+					exts := extractFromProperty(flow, uuid, property, texts, targetTranslation)
+					extracted = append(extracted, exts...)
+				}
+			})
+		}
+	}
+
+	return extracted, nil
+}
+
+func extractFromProperty(flow flows.Flow, uuid uuids.UUID, property string, texts []string, targetTranslation flows.Translation) []*ExtractedText {
+	extracted := make([]*ExtractedText, 0)
+
+	// look up target translation if we have one
+	targets := make([]string, len(texts))
+	if targetTranslation != nil {
+		translation := targetTranslation.GetTextArray(uuid, property)
+		if translation != nil {
+			for t := range targets {
+				if t < len(translation) {
+					targets[t] = translation[t]
+				}
+			}
+		}
+	}
+
+	for t, text := range texts {
+		if text != "" {
+			extracted = append(extracted, &ExtractedText{
+				Locations: []TextLocation{
+					TextLocation{
+						Flow:     flow,
+						UUID:     uuid,
+						Property: property,
+						Index:    t,
+					},
+				},
+				Base:        text,
+				Translation: targets[t],
+				Unique:      false,
+			})
+		}
+	}
+
+	return extracted
+}
+
+func mergeExtracted(extracted []*ExtractedText) []*ExtractedText {
+	// organize extracted texts by their base text
+	byBase := make(map[string][]*ExtractedText)
+	for _, e := range extracted {
+		byBase[e.Base] = append(byBase[e.Base], e)
+	}
+
+	// get the list of unique base text values and sort A-Z
+	bases := make([]string, 0, len(byBase))
+	for b := range byBase {
+		bases = append(bases, b)
+	}
+	sort.Strings(bases)
+
+	merged := make([]*ExtractedText, 0)
+
+	for _, base := range bases {
+		extractionsForBase := byBase[base]
+
+		differingTranslations := false
+		singleTranslation := extractionsForBase[0].Translation
+		for _, ext := range extractionsForBase[1:] {
+			if singleTranslation == "" {
+				singleTranslation = ext.Translation
+			}
+			if ext.Translation != "" && ext.Translation != singleTranslation {
+				differingTranslations = true
+			}
+		}
+
+		if differingTranslations {
+			// we have differing translations, keep extractions for each location separate
+			for _, e := range extractionsForBase {
+				merged = append(merged, e)
+			}
+		} else {
+			// all translations were the same, create merged extraction for all locations
+			locations := make([]TextLocation, len(extractionsForBase))
+			for i, ext := range extractionsForBase {
+				locations[i] = ext.Locations[0]
+			}
+
+			merged = append(merged, &ExtractedText{
+				Locations:   locations,
+				Base:        base,
+				Translation: singleTranslation,
+				Unique:      true,
+			})
+		}
+	}
+
+	return merged
+}
+
+func createPO(lang envs.Language, extracted []*ExtractedText) *PO {
+	pot := NewPO("Generated by flowxgettext", dates.Now(), lang.ToISO639_2(envs.NilCountry))
+
+	for _, ext := range extracted {
+		context := ""
+		if !ext.Unique {
+			context = fmt.Sprintf("%s/%s:%d", string(ext.Locations[0].UUID), ext.Locations[0].Property, ext.Locations[0].Index)
+		}
+
+		references := make([]string, len(ext.Locations))
+		for i, loc := range ext.Locations {
+			references[i] = fmt.Sprintf("%s/%s/%s:%d", loc.Flow.UUID(), string(loc.UUID), loc.Property, loc.Index)
+		}
+		sort.Strings(references)
+
+		entry := &Entry{
+			Comment: Comment{
+				References: references,
+			},
+			MsgContext: context,
+			MsgID:      ext.Base,
+			MsgStr:     ext.Translation,
+		}
+
+		pot.AddEntry(entry)
+	}
+
+	return pot
+}
